+++ Editor/Core/BuildUploaderWelcomeWindow.cs
@@ -1,4 +1,6 @@
 ﻿using System.Collections.Generic;
+using System.IO;
+using System.Linq;
 using UnityEditor;
 using UnityEngine;
 using Object = UnityEngine.Object;
@@ -20,7 +22,9 @@ namespace Wireframe {
             }
         }
 
-        private GUIStyle headerStyle;
+        private GUIStyle headerLabelStyle;
+        private GUIStyle sectionLabelStyle;
+        private GUIStyle sectionFoldoutStyle;
         private Vector2 scrollPosition;
         
         private List<VersionData> parsedChangeLog;
@@ -32,7 +36,7 @@ namespace Wireframe {
             window.titleContent = new GUIContent("Welcome to Build Uploader!", Utils.WindowIcon);
             
             Rect windowPosition = window.position;
-            windowPosition.size = new Vector2(Screen.currentResolution.width * 0.5f, Screen.currentResolution.height * 0.5f);
+            windowPosition.size = new Vector2(Screen.currentResolution.width * 0.5f, Screen.currentResolution.height * 0.75f);
             windowPosition.center = new Rect(0f, 0f, Screen.currentResolution.width, Screen.currentResolution.height).center;
             window.position = windowPosition; 
             window.Show();
@@ -41,20 +45,143 @@ namespace Wireframe {
         private void OnGUI()
         {
             Parse();
+            DrawLinks();
+            scrollPosition = EditorGUILayout.BeginScrollView(scrollPosition);
             
-            GUILayout.Label("Need help setting up the Build Uploader?");
-
-            GUIStyle style = GUI.skin.label;
-            style.wordWrap = true;
-            GUILayout.Label("Check out the Documentation for a step by step guide on how to set up the Build Uploader.", style);
-
-            Links();
+            GUILayout.Label(Utils.WindowLargeIcon, headerLabelStyle);
+            GUILayout.Label("Build Uploader", headerLabelStyle);
+            
+            GUILayout.Label("Welcome to the Build Uploader!");
+            GUILayout.Label("This tool is designed to make it easy to make a build and upload it to all kinds of services.");
+            
+            
+            GUILayout.Label("- Want more information? See the Documentation!");
+            GUILayout.Label("- Want to talk to the Dev or others that use the Build Uploader? Join our Discord!");
+            GUILayout.Label("- Want to see the source code or view in progress changes/fixes? Go to Github!");
+            GUILayout.Label("- Want to ask questions or report a bug or suggest changes? Report Bug/Suggest Feature!");
+            GUILayout.Label("- Want to support the project? Check it out on the Unity Asset Store or press Support Me!");
+            
+            EditorGUILayout.Space(20);
+            
+            DrawSetupCheckList();
 
             EditorGUILayout.Space();
-            GUILayout.Label("Changelog");
+            
+            GUILayout.Label("Changelog", sectionLabelStyle);
             using (new EditorGUILayout.VerticalScope("box"))
             {
-                Changes();
+                DrawChanges();
+            }
+            
+            EditorGUILayout.EndScrollView();
+        }
+
+        private void DrawSetupCheckList()
+        {
+            bool oneServiceReadyToBuild = InternalUtils.AllServices().Any(a => a.IsReadyToStartBuild(out _));
+            bool oneServiceProjectSettingsSetup = InternalUtils.AllServices().Any(a =>
+            {
+                bool success = a.IsReadyToStartBuild(out _) && a.IsProjectSettingsSetup();
+                return success;
+            });
+            bool oneUploadProfileSetup = UploadProfilesExist();
+            
+            bool allComplete = oneServiceReadyToBuild && oneServiceProjectSettingsSetup && oneUploadProfileSetup;
+            
+            bool show = EditorPrefs.GetBool("BuildUploader_showHowToSetup", true);
+            bool newShow = EditorGUILayout.Foldout(show, new GUIContent("Setup checklist", SuccessIcon(allComplete, true)), sectionFoldoutStyle);
+            if (newShow != show)
+            {
+                EditorPrefs.SetBool("BuildUploader_showHowToSetup", newShow);
+            }
+            
+
+            GUILayout.Label("Need more help setting up the Build Uploader?");
+            
+            if (newShow)
+            {
+                GUIStyle header = new GUIStyle(GUI.skin.label);
+                header.fontStyle = FontStyle.Bold;
+
+                GUIStyle scopeIndex = new GUIStyle(GUIStyle.none);
+                scopeIndex.margin.left = 10;
+
+                GUIStyle mainScope = "box";
+                mainScope.margin.left = 10;
+                
+                
+                using (new EditorGUILayout.VerticalScope(mainScope))
+                {
+                    GUILayout.Label("Setup Preferences (Edit->Preferences)", header);
+                    GUILayout.Label("These are settings for your project and not shared with anyone.");
+
+                    using (new EditorGUILayout.VerticalScope(scopeIndex))
+                    {
+                        GUILayout.Label($"\nBuild Uploader -> General");
+                        DrawCheckList("Change Cached Builds to a smaller path. eg: C:/CachedBuilds",
+                            null, !Preferences.CacheFolderPath.Equals(Preferences.DefaultCacheFolder));
+
+                        GUILayout.Label($"\nBuild Uploader -> Services");
+                        DrawCheckList("Enable and enter credentials for all services you want to use",
+                            "Enable Steamworks, download and install SteamSDK and enter your username.",
+                            oneServiceReadyToBuild);
+                    }
+
+                    GUILayout.Label("\nSetup Project Settings (Edit->Project Settings)", header);
+                    GUILayout.Label("These are specific to your project and will be shared with anyone with access to your source code.");
+
+                    using (new EditorGUILayout.VerticalScope(scopeIndex))
+                    {
+                        GUILayout.Label($"\nBuild Uploader -> Services");
+                        DrawCheckList("Enter settings for all Services you want to use",
+                            "For Steamworks add a new App for your game and any branches you want to use.",
+                            oneServiceProjectSettingsSetup);
+                    }
+
+                    GUILayout.Label("\nSetup Upload Profile (Window -> Build Uploader -> Open Window)", header);
+                    using (new EditorGUILayout.VerticalScope(scopeIndex))
+                    {
+                        DrawCheckList(
+                            "Create an Upload Config so you can make a build and upload it to a service of your choosing.",
+                            "", oneUploadProfileSetup);
+                    }
+                }
+            }
+        }
+
+        private bool UploadProfilesExist()
+        {
+            string path = WindowUploadTab.UploadProfilePath;
+            if (string.IsNullOrEmpty(path) || !Directory.Exists(path))
+            {
+                return false;
+            }
+            
+            string[] files = Directory.GetFiles(path, "*.json");
+            return files.Length > 0;
+        }
+
+        private Texture2D SuccessIcon(bool success, bool big)
+        {
+            if (big)
+            {
+                return success ? Utils.CheckIcon : Utils.CrossIcon;
+            }
+            return success ? Utils.CheckIconSmall : Utils.CrossIconSmall;
+        }
+        
+        private void DrawCheckList(string text, string example, bool isComplete)
+        {
+            GUIStyle exampleStyle = new GUIStyle(GUI.skin.label);
+            exampleStyle.fontStyle = FontStyle.Italic;
+            exampleStyle.wordWrap = true;
+            exampleStyle.richText = true;
+            exampleStyle.onNormal.textColor = Color.black;
+            
+            GUILayout.Label(new GUIContent(text, SuccessIcon(isComplete, false)));
+            if (!string.IsNullOrEmpty(example))
+            {
+                GUILayout.Label($"\tExample: {example}", exampleStyle);
             }
         }
 
@@ -65,10 +192,22 @@ namespace Wireframe {
                 return;
             }
             
-            headerStyle = new GUIStyle(EditorStyles.boldLabel);
-            headerStyle.fontSize = 24;
+            headerLabelStyle = new GUIStyle(GUI.skin.label);
+            headerLabelStyle.wordWrap = true;
+            headerLabelStyle.alignment = TextAnchor.MiddleCenter;
+            headerLabelStyle.fontStyle = FontStyle.Bold;
+            headerLabelStyle.fontSize = 24;
             
-            var path = "Packages/com.veugeljame.builduploader/CHANGELOG.md";
+            sectionLabelStyle = new GUIStyle(GUI.skin.label);
+            sectionLabelStyle.wordWrap = true;
+            sectionLabelStyle.alignment = TextAnchor.MiddleLeft;
+            sectionLabelStyle.fontStyle = FontStyle.Normal;
+            sectionLabelStyle.fontSize = 18;
+
+            sectionFoldoutStyle = new GUIStyle(EditorStyles.foldout);
+            sectionFoldoutStyle.fontSize = 16;
+            
+            string path = Path.Combine(Utils.s_packagePath, "CHANGELOG.md");
             Object loadAssetAtPath = AssetDatabase.LoadAssetAtPath(path, typeof(TextAsset));
             string allText = loadAssetAtPath is TextAsset textAsset ? textAsset.text : "";
             string[] lines = allText.Split('\n');
@@ -102,12 +241,11 @@ namespace Wireframe {
             }
         }
 
-        private void Changes()
+        private void DrawChanges()
         {
             GUIStyle foldoutStyle = new GUIStyle(EditorStyles.foldout);
             foldoutStyle.fontSize = 18;
             
-            scrollPosition = EditorGUILayout.BeginScrollView(scrollPosition);
             
             for (int i = 0; i < parsedChangeLog.Count; i++)
             {
@@ -132,8 +270,6 @@ namespace Wireframe {
                     GUILayout.Space(10);
                 }
             }
-
-            EditorGUILayout.EndScrollView();
         }
 
         private void Draw(string[] lines)
@@ -163,7 +299,7 @@ namespace Wireframe {
                 {
                     // Header
                     line = line.Substring(1).Trim();
-                    style = headerStyle;
+                    style = headerLabelStyle;
                 }
                 else if (line.Trim().StartsWith("-"))
                 {
@@ -248,33 +384,38 @@ namespace Wireframe {
             }
         }
 
-        private static void Links()
+        private static void DrawLinks()
         {
             using (new EditorGUILayout.HorizontalScope())
             {
-                if (GUILayout.Button("Documentation"))
+                if (GUILayout.Button(new GUIContent("Documentation", Utils.LinkIcon)))
                 {
                     Application.OpenURL("https://github.com/JamesVeug/UnitySteamBuildUploader/wiki");
                 }
                 
-                if (GUILayout.Button("Discord"))
+                if (GUILayout.Button(new GUIContent("Discord", Utils.LinkIcon)))
                 {
                     Application.OpenURL("https://discord.gg/R2UjXB6pQ8");
                 }
                 
-                if (GUILayout.Button("Github"))
+                if (GUILayout.Button(new GUIContent("Github", Utils.LinkIcon)))
                 {
                     Application.OpenURL("https://github.com/JamesVeug/UnitySteamBuildUploader");
                 }
                 
-                if (GUILayout.Button("Support Me"))
+                if (GUILayout.Button(new GUIContent("Asset Store", Utils.LinkIcon)))
                 {
-                    Application.OpenURL("https://buymeacoffee.com/jamesgamesnz");
+                    Application.OpenURL("https://assetstore.unity.com/packages/tools/utilities/build-uploader-306907");
                 }
                 
-                if (GUILayout.Button("Report Bug"))
+                if (GUILayout.Button(new GUIContent("Report Bug / Suggest Feature", Utils.LinkIcon)))
                 {
                     Application.OpenURL("https://github.com/JamesVeug/UnitySteamBuildUploader/issues");
+                }
+                
+                if (GUILayout.Button(new GUIContent("Support Me", Utils.LinkIcon)))
+                {
+                    Application.OpenURL("https://buymeacoffee.com/jamesgamesnz");
                 }
             }
         }