+++ Editor/Services/AService.cs
@@ -6,9 +6,10 @@ namespace Wireframe
     public abstract class AService
     {
         public abstract string ServiceName { get; }
-        public abstract string[] SearchKeyworks { get; }
+        public abstract string[] SearchKeywords { get; }
         internal virtual WindowTab WindowTabType => null;
         public abstract bool IsReadyToStartBuild(out string reason);
+        public abstract bool IsProjectSettingsSetup();
         public abstract void PreferencesGUI();
         public abstract void ProjectSettingsGUI();
         public virtual bool HasProjectSettingsGUI => false;
@@ -20,7 +21,7 @@ namespace Wireframe
                 return true;
             }
 
-            bool matchesSearchKeywords = SearchKeyworks.Any(a => Utils.Contains(a, search, StringComparison.OrdinalIgnoreCase));
+            bool matchesSearchKeywords = SearchKeywords.Any(a => Utils.Contains(a, search, StringComparison.OrdinalIgnoreCase));
             return matchesSearchKeywords;
         }
     }