+++ Editor/Services/DiscordService/DiscordService.cs
@@ -6,7 +6,7 @@ namespace Wireframe
     internal partial class DiscordService : AService
     {
         public override string ServiceName => "Discord";
-        public override string[] SearchKeyworks => new string[]{"discord", "chat", "messaging"};
+        public override string[] SearchKeywords => new string[]{"discord", "chat", "messaging"};
         
         public DiscordService()
         {
@@ -23,6 +23,17 @@ namespace Wireframe
 
             reason = "";
             return true;
+        }
+
+        public override bool IsProjectSettingsSetup()
+        {
+            DiscordConfig data = DiscordUIUtils.GetConfig(false);
+            if (data == null)
+            {
+                return false;
+            }
+            
+            return data.servers.Count > 0;
         }
     }
 }