+++ Editor/Services/GithubService/GithubService.cs
@@ -3,7 +3,7 @@
     internal partial class GithubService : AService
     {
         public override string ServiceName => "Github";
-        public override string[] SearchKeyworks => new string[]{"Git", "hub"};
+        public override string[] SearchKeywords => new string[]{"Git", "hub"};
 
         public GithubService()
         {
@@ -25,6 +25,11 @@
             }
 
             reason = "";
+            return true;
+        }
+
+        public override bool IsProjectSettingsSetup()
+        {
             return true;
         }