+++ Editor/Services/SlackService/SlackService.Preferences_Provider.cs
@@ -13,7 +13,7 @@ namespace Wireframe
                 new SlackService_Preferences_Provider("Preferences/Build Uploader/Services/Slack", SettingsScope.User)
                 {
                     label = "Slack",
-                    keywords = InternalUtils.AllServices().FirstOrDefault(a=>a is SlackService).SearchKeyworks
+                    keywords = InternalUtils.AllServices().FirstOrDefault(a=>a is SlackService).SearchKeywords
                 };
             return provider;
         }