+++ Editor/Services/UnityCloudService/UnityCloudService.cs
@@ -6,7 +6,7 @@
     internal partial class UnityCloudService : AService
     {
         public override string ServiceName => "Unity Cloud";
-        public override string[] SearchKeyworks => new string[]{"unity", "cloud", "unity cloud", "ci", "devops"};
+        public override string[] SearchKeywords => new string[]{"unity", "cloud", "unity cloud", "ci", "devops"};
 
         public UnityCloudService()
         {
@@ -41,6 +41,11 @@
             
             
             reason = "";
+            return true;
+        }
+
+        public override bool IsProjectSettingsSetup()
+        {
             return true;
         }